+++ src/s3/s3-gateway.ts
@@ -234,16 +234,28 @@ export class S3TestStore implements bs.TestGateway {
   }
 }
 
-export function registerS3StoreProtocol(protocol = "s3:", overrideBaseURL?: string) {
-  return bs.registerStoreProtocol({
-    protocol,
-    overrideBaseURL,
-    gateway: async (logger) => {
-      return new S3Gateway(logger);
-    },
-    test: async (sthis: SuperThis) => {
-      const gateway = new S3Gateway(sthis);
-      return new S3TestStore(sthis, gateway);
-    },
-  });
+export interface versionUnregister {
+  (): void;
+  readonly version: string;
+}
+
+export function registerS3StoreProtocol(protocol = "s3:", overrideBaseURL?: string): versionUnregister {
+  URI.protocolHasHostpart(protocol);
+  const unreg: versionUnregister = (() => {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const _f: any = bs.registerStoreProtocol({
+      protocol,
+      overrideBaseURL,
+      gateway: async (sthis) => {
+        return new S3Gateway(sthis);
+      },
+      test: async (sthis: SuperThis) => {
+        const gateway = new S3Gateway(sthis);
+        return new S3TestStore(sthis, gateway);
+      },
+    });
+    _f.version = S3_VERSION;
+    return _f;
+  })();
+  return unreg;
 }
